requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
google-genai>=1.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
